Moved CfgWeapons/CfgVehicles to own files
